Support user-provided fixers
Avoid erroring out when a non-standard fixer is specified, and instead
hand it over to the refactoring tool so it can try to load it.

While at it, add a command-line option to add the current directory
to sys.path, to make it easier for users to let their fixers be found.